Format output with go/printer
This turns something like

    type honnef.co/go/cups/raster.Header struct{MediaClass string;
    MediaColor string; MediaType string; OutputType string; AdvanceDistance
    int; AdvanceMedia int; Collate bool; CutMedia int; Duplex bool}

into

    import "honnef.co/go/cups/raster"

    type Header struct {
    	MediaClass string
    	MediaColor string
    	// Foo.
    	MediaType       string
    	OutputType      string
    	AdvanceDistance int
    	AdvanceMedia    int
    	Collate         bool
    	CutMedia        int
    	Duplex          bool
    }

which is much more readable, and in line with what 'go doc' outputs.

Because the type name doesn't include the import path anymore, we print
an explicit 'import' line.

A side effect of this change is that the title line doesn't have a
trailing space anymore, which would be highlighted in some editors.