Don't crash if file failed to parse completely
Sometimes, even go/parser can't recover. In that case, we end up with an
incomplete file set. It might be nice to specifically detect this case
and provide a more useful error, but for now it's good enough not to
crash and just complain that the file wasn't found in the package.