Remove ast.File doc
This change prevents us for showing doc for *ast.File.

As a result, we no longer show doc when the cursor is over the package clause, but as pointed out in #23, when you're working in a package you usually don't need the package doc for it.

The benefit of the change is that we will show "no documentation found" instead of the package doc when the cursor is over whitespace elsewhere in the file.

Fixes #23