Feat: Add local package support
The current version of gogetdoc can't work for a local package,
This means if your code is not put in GOPATH, then you can't get any
help with gogetdoc.

This commit changes the current behavior and adds a function named
`importLocalPackage` which can import local packages, when
`buildutil.ContainingPackage(...)` failed, `importLocalPackage` will
be called instead of return an error.

It also adds two test cases for test `Run` from outside of GOPATH.