Vendor dependencies
This project uses a go.mod file to pin dependencies to specific commits,
but some users of Go may not be using go modules yet since it's an
optional feature in Go 1.11. Vendoring dependencies will allow this
package to build predictably with and without go modules for users of
older versions of Go and for users of Go 1.11 who have go modules
disabled.